+++ CHANGELOG.md
@@ -10,6 +10,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 > *Note*: Some version numbers may be skipped to align feature set with the Node.js version.
 
 ---
+## Version 4.2.40, 5/9/2025
+
+### Added
+
+1. support of "flows" in the "modules.autostart" feature
+2. "length" type matching feature
+3. pipeline for-loop handles model variable in comparator's right hand side
+
+### Removed
+
+N/A
+
+### Changed
+
+N/A
+
+---
 ## Version 4.2.39, 5/4/2025
 
 ### Added